requirements.txt
Django==1.5.1
Markdown==2.3.1
django-autoslug==1.7.1
django-extensions==1.1.1
django-filter==0.6
django-model-utils==1.4.0
djangorestframework==2.3.6
docutils==0.11
-e /Users/guglielmo/Workspace/django-popolo
